server/requirements.txt
fastapi
uvicorn
mangum
pydantic
python-dotenv
langchain
langchain-aws
langchain-core
langgraph
boto3
tzdata
